Move downloadDatastream() to DatastreamTable.vue
